refactor: extract tests to sibling files for 9 more god-files
Mechanical test-sibling extraction. No behavior change.

  tools/impl/browser/browser.rs           2938 -> 2054 (-884)
  tokenjuice/reduce.rs                    2376 ->  907 (-1469)
  agent/harness/session/turn.rs           2003 -> 1435 (-568)
  agent/harness/subagent_runner/ops.rs    1725 -> 1146 (-579)
  channels/providers/lark.rs              1513 ->  900 (-613)
  core/jsonrpc.rs                         1488 -> 1110 (-378)
  channels/controllers/ops.rs             1413 -> 1051 (-362)
  voice/server.rs                         1398 -> 1069 (-329)
  channels/providers/whatsapp_storage.rs  1345 -> 1264 (-81)

Also rolls in rustfmt fixes for agent/prompts/{mod,sections/*}.rs
from an earlier commit that missed them.

Same pattern as 825d734b / 92da9ccb / ef411cec.
