+++ src/openhuman/agent/harness/tool_loop_tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 use crate::openhuman::approval::ApprovalManager;
 use crate::openhuman::config::AutonomyConfig;
