+++ src/openhuman/agent/prompts/sections/mod.rs
@@ -7,14 +7,14 @@ mod user_files;
 mod user_memory;
 mod workspace;
 
-pub use datetime::{DateTimeSection, render_datetime};
-pub use identity::{IdentitySection, render_identity};
-pub use runtime::{RuntimeSection, render_runtime};
-pub use safety::{SafetySection, render_safety};
-pub use tools::{ToolsSection, render_tools};
-pub use user_files::{UserFilesSection, render_user_files};
-pub use user_memory::{UserMemorySection, render_user_memory};
-pub use workspace::{WorkspaceSection, render_workspace};
+pub use datetime::{render_datetime, DateTimeSection};
+pub use identity::{render_identity, IdentitySection};
+pub use runtime::{render_runtime, RuntimeSection};
+pub use safety::{render_safety, SafetySection};
+pub use tools::{render_tools, ToolsSection};
+pub use user_files::{render_user_files, UserFilesSection};
+pub use user_memory::{render_user_memory, UserMemorySection};
+pub use workspace::{render_workspace, WorkspaceSection};
 
 pub(crate) use tools::render_pformat_signature_for_box_tool;
