+++ src/openhuman/config/ops_tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 use tempfile::tempdir;
