+++ src/openhuman/memory/ingestion_tests.rs
@@ -1,4 +1,3 @@
-
 use std::sync::Arc;
 
 use serde_json::json;
