+++ src/openhuman/memory/ops_tests.rs
@@ -1,4 +1,3 @@
-
 use serde_json::json;
 
 use super::{build_retrieval_context, filter_hits_by_document_ids, format_llm_context_message};
